+++ src/CoordinationIdFactory.php
@@ -5,16 +5,8 @@ namespace byrokrat\id;
 /**
  * Create coordination id objects from raw id string
  */
-class CoordinationIdFactory extends IdFactory
+class CoordinationIdFactory extends AbstractFactoryDecorator
 {
-    use Component\Factory;
-
-    /**
-     * Instantiate ID object
-     *
-     * @param  string $raw Raw id string
-     * @return IdInterface
-     */
     protected function createNewInstance($raw)
     {
         return new CoordinationId($raw);