+++ src/IdFactoryInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace byrokrat\id;
+
+/**
+ * The base id factory interface
+ */
+interface IdFactoryInterface
+{
+    /**
+     * Create ID object from raw id string
+     *
+     * @param  string $raw Raw id string
+     * @return IdInterface
+     */
+    public function createId($raw);
+}