+++ src/OrganizationIdFactory.php
@@ -5,16 +5,8 @@ namespace byrokrat\id;
 /**
  * Create organization id objects from raw id string
  */
-class OrganizationIdFactory extends IdFactory
+class OrganizationIdFactory extends AbstractFactoryDecorator
 {
-    use Component\Factory;
-
-    /**
-     * Instantiate ID object
-     *
-     * @param  string $raw Raw id string
-     * @return IdInterface
-     */
     protected function createNewInstance($raw)
     {
         return new OrganizationId($raw);