+++ src/PersonalIdFactory.php
@@ -5,16 +5,8 @@ namespace byrokrat\id;
 /**
  * Create personal id objects from raw id string
  */
-class PersonalIdFactory extends IdFactory
+class PersonalIdFactory extends AbstractFactoryDecorator
 {
-    use Component\Factory;
-
-    /**
-     * Instantiate ID object
-     *
-     * @param  string $raw Raw id string
-     * @return IdInterface
-     */
     protected function createNewInstance($raw)
     {
         return new PersonalId($raw);